Issue #48: Add dummy data to the database for testing
Added seeding with inserting 10 dummy users and a preset admin.
The login credentials for the admin is 'root@eotvos.elte.hu' and 'asdasdasd'.
To run the seeding with a refreshed migration give the command:
php artisan migrate:refresh --seed
Before that, you might have to give the command:
composer dump-autoload