Upgrade free page handling
This commit adds some fixes, design and enhancements.

Fixes #122, which adds deadline for free pages.
Fixes #97, which adds free pages with different deadlines.
Fixes #149, where free pages were not accounted for when printing.
Partially fixes #147, free pages are listed now for admins. A different
page for listing balances is left.

Thanks for the effort, if you review this one.