Enable VS code dev containers (#623)
* Enable VS code dev containers

The VS code dev container allows us to develop in docker containers the
same as in local environments. It's similar to remote SSH setups, but
here we are inside the container instead of a remote machine. The huge
advantage of using this is developers will be able to start working on
the project right after opening it, no need to worry about the
complicated setup requirements.

For details see:
https://code.visualstudio.com/docs/remote/containers

The devcontainer.json file tells VS code how to create the dev
container. The base of this is a docker compose file, which I have
created with Laravel Sail. This will build the Laravel application,
downloads all the necessary components (using composer and npm install),
and also start a mysql service.

I also added a script that will install some tools and customization,
migrates and seeds the database, and compiles the frontend.

I've added some extensions to run in the container, but feel free to add
anything else you're using.

Requirements:
 - You need to install Docker (and WSL2 on Windows)
 - You need to install VS code
 - You need to install the Remote Development extension pack in VS code.

For more details see:
https://code.visualstudio.com/docs/remote/containers#_system-requirements

* Update docs

* Add defaults for WWWGROUP and WWWUSER

* change to php 7.4